Solution #2 in section 4.1 may not work
In my testing, using an FQDN name for the conditional forwarding on premises required to use the privatelink name, not the real FQDN.  In my testing this configuration (conditional forwarding set to real FQDN, which forwarded to custom DNS server in Azure) got the correct private response the first time, and then the public address every query after.  If the privatelink name is used for the conditional forwarder, then every query works as intended.